Somewhat functioning swap command

+++ src/Library.h
@@ -7,6 +7,6 @@ int isNumberStr(char *c);
 int strtoint(char *str);
 long strtolong(char *str);
 int intsize(int x);
-long longsize(long x);
+int longsize(long x);
 
 #endif //COURSEWORK2_LIBRARY_H
